backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.25.0
azure-storage-blob>=12.19.0
aiohttp>=3.9.0
sse-starlette>=1.8.0
